feat(batch): refactor logic
