+++ src/orchestration/operators/batch/__init__.py
@@ -28,181 +28,10 @@ BatchIndexOperator -> ManifestGenerator -> BatchIndex -
 ```
 """
 
-from __future__ import annotations
+from orchestration.operators.batch.batch_index_operator import BatchIndexOperator
+from orchestration.operators.batch.batch_job_operator import BatchJobOperator
 
-import logging
-import time
-from functools import cached_property
-
-from airflow.exceptions import AirflowSkipException
-from airflow.models.baseoperator import BaseOperator
-from airflow.providers.google.cloud.operators.cloud_batch import CloudBatchSubmitJobOperator
-from pydantic import BaseModel
-
-from orchestration.models.batch import BatchIndexOperatorSpec, BatchJobOperatorSpec
-from orchestration.models.batch.environment import EnvironmentRegistrySpec
-from orchestration.operators.batch.manifest_generators import MANIFEST_GENERATOR_MAP
-from orchestration.operators.batch.manifest_generators.proto import ProtoManifestGenerator
-from orchestration.utils.common import GCP_PROJECT_GENETICS, GCP_REGION
-
-logger = logging.getLogger(__name__)
-
-
-class BatchIndexRow(BaseModel):
-    """Representation of a single row in the batch index. Each row corresponds to a single batch job.
-
-    a Pydantic model representing a single batch job, carrying a
-    zero-based index and the `EnvironmentRegistrySpec` (one environment per task)
-    for that job.
-    """
-
-    idx: int
-    """Index of the batch job. This can be used to create unique identifiers for batch jobs and their tasks."""
-    environments: EnvironmentRegistrySpec
-    """Environment specification for the batch job. This will be used to create the Environment object for each task."""
-
-
-class BatchIndex:
-    """In-memory index of batch jobs produced by partitioning an EnvironmentRegistrySpec.
-
-    Intended usage is a two-step sequence:
-        1. Call `partition(max_task_count)` to split the registry into chunks, one per batch job.
-        2. Access `rows` to get the resulting list of `BatchIndexRow` objects.
-
-    `BatchIndex`: wraps an `EnvironmentRegistrySpec` and partitions it into chunks
-    that each become a separate batch job. The intended call sequence is::
-
-        index = generator.generate_batch_index()
-        rows = index.partition(max_task_count).rows
-
-    ``partition`` splits the registry into `EnvironmentRegistrySpec` chunks capped
-    at ``max_task_count`` tasks each. ``rows`` converts those chunks into
-    `BatchIndexRow` objects consumed by `BatchJobOperator`.
-    """
-
-    def __init__(
-        self,
-        env_registry: EnvironmentRegistrySpec,
-    ) -> None:
-        """Initialise a BatchIndex from an environment registry.
-
-        Args:
-            env_registry (EnvironmentRegistrySpec): Registry of environment specifications,
-                one per task to be distributed across batch jobs.
-        """
-        self.environment_registry = env_registry
-        self.env_batches: list[EnvironmentRegistrySpec] = []
-
-    def partition(self, max_task_count: int) -> BatchIndex:
-        """Partition the environment registry into chunks, each capped at `max_task_count` tasks.
-
-        Args:
-            max_task_count (int): Maximum number of tasks (environments) allowed per batch job.
-
-        Returns:
-            BatchIndex: self, enabling method chaining (e.g. ``index.partition(n).rows``).
-
-        Raises:
-            AirflowSkipException: If the environment registry is empty.
-        """
-        if self.environment_registry.empty:
-            msg = "BatchIndex can not partition variable list, as list is empty."
-            logger.warning(msg)
-            raise AirflowSkipException(msg)
-
-        self.env_batches = self.environment_registry.partition(max_task_count=max_task_count)
-        logger.info("Created %s task list batches.", len(self.env_batches))
-
-        return self
-
-    @cached_property
-    def rows(self) -> list[BatchIndexRow]:
-        """Return one BatchIndexRow per partitioned batch job.
-
-        Each row carries a zero-based index and the EnvironmentRegistrySpec for that job's tasks.
-        Requires ``partition()`` to have been called first.
-
-        Returns:
-            list[BatchIndexRow]: Ordered list of batch job descriptors.
-
-        Raises:
-            AirflowSkipException: If no rows are available (e.g. ``partition()`` was never called).
-        """
-        rows: list[BatchIndexRow] = []
-        logger.info("Preparing BatchIndexRows. Each row represents a batch job.")
-        for idx, batch in enumerate(self.env_batches):
-            rows.append(BatchIndexRow(idx=idx, environments=batch))
-
-        logger.info("Prepared %s BatchIndexRows", len(rows))
-        if not rows:
-            raise AirflowSkipException("Empty BatchIndexRows will not allow to create batch task. Skipping downstream")
-        return rows
-
-    def __repr__(self) -> str:
-        """Get batch index string representation."""
-        return f"BatchIndex(env_registry={self.environment_registry})"
-
-
-class BatchIndexOperator(BaseOperator):
-    """Operator to prepare google batch job index and partition it into the manifests.
-
-    Each manifest prepared by the operator should create an environment for a single batch job.
-    Each row of the individual manifest should represent individual batch task.
-    """
-
-    def __init__(
-        self,
-        batch_index_specs: BatchIndexOperatorSpec,
-        **kwargs,
-    ) -> None:
-        self.manifest_generator = self.get_generator(batch_index_specs.pointer)
-        self.generator_specs = batch_index_specs.generator_specs
-        self.max_task_count = batch_index_specs.max_task_count
-        super().__init__(**kwargs)
-
-    @classmethod
-    def get_generator(cls, label: str) -> type[ProtoManifestGenerator]:
-        """Get the generator by it's label in the registry."""
-        try:
-            return MANIFEST_GENERATOR_MAP[label]
-        except KeyError:
-            raise KeyError(f"Manifest generator with label {label} not found in the manifest generator registry.")
-
-    def execute(self, **kwargs) -> list[BatchIndexRow]:
-        """Execute the operator."""
-        generator = self.manifest_generator.from_generator_config(self.generator_specs)
-        index = generator.generate_batch_index()
-        if not self.max_task_count:
-            # if specified 0 or not specified in the config, then assume to use the number
-            # of tasks that is in the output of the BatchIndex.environment_registry from manifest generation
-            self.max_task_count = len(index.environment_registry)
-        self.log.info(index)
-        partitioned_index = index.partition(self.max_task_count)
-        return partitioned_index.rows
-
-
-class BatchJobOperator(CloudBatchSubmitJobOperator):
-    """Generic Batch Job operator.
-
-    This operator has to be used in conjunction to the BatchIndexOperator.
-    It runs the google batch jobs defined by the BatchIndexOperator.
-    """
-
-    def __init__(
-        self,
-        job_name: str,
-        batch_index_row: BatchIndexRow,
-        batch_job_specs: BatchJobOperatorSpec,
-        project_id: str = GCP_PROJECT_GENETICS,
-        region: str = GCP_REGION,
-        **kwargs,
-    ):
-        job = batch_job_specs.job.build(task_environments=batch_index_row.environments)
-        super().__init__(
-            project_id=project_id,
-            region=region,
-            job_name=f"{job_name}-job-{batch_index_row.idx}-{time.strftime('%Y%m%d-%H%M%S')}",
-            job=job,
-            deferrable=False,
-            **kwargs,
-        )
+__all__ = [
+    "BatchIndexOperator",
+    "BatchJobOperator",
+]
